python/requirements.txt
PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
glfw>=2.7.0
numpy>=1.26.0
